requirements.txt
azure-core==1.23.0
azure-cosmos==4.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.1.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
